fix: resolve MEDIUM audit items (M1, M3, M4, M5) and BigInt dashboard bug
M1: Replace native <img> with next/image in chat-header and contact-profile,
    add remotePatterns config for WhatsApp profile pics
M3: Add lockDuration, maxStalledCount, stalledInterval, removeOnComplete,
    removeOnFail to all BullMQ workers (chat + ERP)
M4: Dynamic imports for Recharts chart components in dashboard
M5: Split channel-qr-dialog.tsx (401 lines) into 3 files:
    channel-qr-dialog.tsx (parent), channel-qr-tab.tsx, channel-pairing-tab.tsx

Also fixes BigInt serialization error in dashboard stats endpoint by casting
COUNT(*)::int in the raw SQL query.

Deferred M2, M6, M7 to docs/plans/backlog-medium-features.md (require new
backend routes).

+++ apps/server/src/routes/v1/stats-routes.ts
@@ -1,7 +1,7 @@
-import type { FastifyInstance, FastifyRequest, FastifyReply } from 'fastify'
 import { prisma } from '@repo/db'
-import { tenantMiddleware } from '../../middlewares/tenant-middleware.js'
+import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify'
 import { requireAbility } from '../../middlewares/ability-middleware.js'
+import { tenantMiddleware } from '../../middlewares/tenant-middleware.js'
 import { dashboardStatsQuerySchema } from '../../schemas/stats.schemas.js'
 
 export async function statsRoutes(app: FastifyInstance) {
@@ -99,8 +99,8 @@ export async function statsRoutes(app: FastifyInstance) {
         prisma.$queryRaw`
           SELECT
             TO_CHAR(DATE_TRUNC('month', "createdAt"), 'YYYY-MM') as month,
-            COUNT(*) FILTER (WHERE "stage" != 'LOST') as proposals,
-            COUNT(*) FILTER (WHERE "stage" = 'POLICY_ISSUED') as issued
+            COUNT(*) FILTER (WHERE "stage" != 'LOST')::int as proposals,
+            COUNT(*) FILTER (WHERE "stage" = 'POLICY_ISSUED')::int as issued
           FROM "Proposal"
           WHERE "organizationId" = ${orgId}
             AND "createdAt" >= ${cutoffDate}
